+++ TimberCommons/package/CHANGELOG.md
@@ -1,5 +1,6 @@
 # v1.6 (started on 2/7/2024):
 * [Fix #33] Water valve doesn't handle contamination correctly.
+* Make WaterValve pausable.
 
 # v1.5 (3 Feb 2024):
 * Relase irrigation towers code.